pyproject.toml
[project]
name = "ai-news"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.82.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
